Moving NeutronicsPlugin to its own file
